move reuasble code to new parent class; 1st pass
- new class OpenMCOperator to contain shared bits of code between the
  currently existing Operator class and the as-of-yet to be created
  FluxDepletionOperator class
- Operator is now a subclass of OpenMCOperator
- move _distribute method to OpenMCOperator
- move _get_burnable_mats, _extract_number, _set_number_from_mat,
  _set_number_from_results to OpenMCOperator
- split initial condition into openmc.lib dependent and non-dependent
  parts; dependent part goes to Operator, non-dependent part goes to
  OpenMCOperator
- move _update_materials to OpenMCOperator
- move _get_tally_nuclides to OpenMCOperator
- rename _unpack_tallies_and_normalize to _calculate_reaction_rates;
  move to OpenMCOperator
- move get_results_info to OpenMCOperator
