+++ openmc/deplete/__init__.py
@@ -7,6 +7,7 @@ A depletion front-end tool.
 
 from .nuclide import *
 from .chain import *
+from .openmc_operator import *
 from .operator import *
 from .reaction_rates import *
 from .atom_number import *
